Implmented sessionUser middleware provided by users.controller

+++ server/users/users.controller.js
@@ -1,6 +1,7 @@
 const userService = require('./users.service');
 
 module.exports = {
+    attachSessionUser,
     authenticate,
     register,
     getById,
@@ -9,6 +10,18 @@ module.exports = {
     update,
     delete: _delete
 };
+
+function attachSessionUser(req, res, next) {
+    userService.getById(req.user.sub)
+        .then(function (user) {
+            if (user){
+                req.sessionUser = user;
+                next();
+            }
+            else res.status(409).json({ message: 'Session user no longer exists'});
+        })
+        .catch(err => next(err));
+}
 
 function authenticate(req, res, next) {
     userService.authenticate(req.body)
@@ -35,19 +48,23 @@ function getAll(req, res, next) {
 }
 
 function getCurrent(req, res, next) {
-    userService.getById(req.user.sub)
-        .then(user => user ? res.json(user) : res.sendStatus(404))
-        .catch(err => next(err));
+    req.sessionUser ? res.json(req.sessionUser) : res.sendStatus(404);
 }
 
 function update(req, res, next) {
-    userService.update(req.params.id, req.body)
-        .then(() => res.json({}))
-        .catch(err => next(err));
+    if (req.sessionUser.id !== req.params.id)
+        res.sendStatus(404);
+    else
+        userService.update(req.params.id, req.body)
+            .then(() => res.json({}))
+            .catch(err => next(err));
 }
 
 function _delete(req, res, next) {
-    userService.delete(req.params.id)
-        .then(() => res.json({}))
-        .catch(err => next(err));
+    if (req.sessionUser.id !== req.params.id)
+        res.sendStatus(404);
+    else
+        userService.delete(req.params.id)
+            .then(() => res.json({}))
+            .catch(err => next(err));
 }
